requirements.txt
PyPDF2
python-dotenv
requests
streamlit
openai
